requirements.txt
bottle~=0.12.21
sqlobject~=3.9.1
eyed3
python-dateutil~=2.8.2